+++ services/pom.xml
@@ -112,6 +112,16 @@
             <groupId>commons-collections</groupId>
             <artifactId>commons-collections</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.apache.karaf.cellar</groupId>
+            <artifactId>org.apache.karaf.cellar.core</artifactId>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.karaf.cellar</groupId>
+            <artifactId>org.apache.karaf.cellar.config</artifactId>
+            <scope>provided</scope>
+        </dependency>
     </dependencies>
 
     <build>
@@ -155,6 +165,18 @@
                                     </file>
                                     <type>cfg</type>
                                     <classifier>thirdpartycfg</classifier>
+                                </artifact>
+                                <artifact>
+                                    <file>
+                                        src/main/resources/org.apache.unomi.cluster.cfg
+                                    </file>
+                                    <type>cfg</type>
+                                    <classifier>clustercfg</classifier>
+                                </artifact>
+                                <artifact>
+                                    <file>src/main/resources/hazelcast.xml</file>
+                                    <type>xml</type>
+                                    <classifier>hazelcastconfig</classifier>
                                 </artifact>
                             </artifacts>
                         </configuration>
