+++ services/src/main/resources/org.apache.unomi.cluster.cfg
@@ -14,24 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-cluster.name=contextElasticSearch
-index.name=context
-monthlyIndex.numberOfShards=3
-monthlyIndex.numberOfReplicas=0
-numberOfShards=5
-numberOfReplicas=0
-defaultQueryLimit=10
-
-# The following settings control the behavior of the BulkProcessor API. You can find more information about these
-# settings and their behavior here : https://www.elastic.co/guide/en/elasticsearch/client/java-api/2.4/java-docs-bulk-processor.html
-# The values used here are the default values of the API
-bulkProcessor.name=unomi-bulk
-bulkProcessor.concurrentRequests=1
-bulkProcessor.bulkActions=1000
-bulkProcessor.bulkSize=5MB
-bulkProcessor.flushInterval=5s
-bulkProcessor.backoffPolicy=exponential
-
-minimalElasticSearchVersion=5.0.0
-maximalElasticSearchVersion=5.2.0
+cluster.group=default
+cluster.jmxUsername=karaf
+cluster.jmxPassword=karaf
+cluster.jmxPort=1099
