+++ services/src/main/resources/org.apache.unomi.services.cfg
@@ -19,7 +19,7 @@
 profile.purge.interval=1
 
 # Purge profiles that have been inactive for a specific number of days
-profile.purge.inactiveTime=30
+profile.purge.inactiveTime=180
 
 # Purge profiles that have been created for a specific number of days
 profile.purge.existTime=-1
